refactor(core): rename and move baseprimitve to hook
