+++ @navikt/core/react/src/primitives/grid/HGrid.tsx
@@ -1,20 +1,20 @@
 import React, { forwardRef } from "react";
 import { type OverridableComponent, omit } from "../../utils-external";
+import { Slot } from "../../utils/components/slot/Slot";
+import { cl } from "../../utils/helpers";
+import {
+  getResponsiveProps,
+  getResponsiveValue,
+} from "../../utils/hooks/base-primitive/helpers/css";
+import type {
+  ResponsiveProp,
+  SpacingScale,
+} from "../../utils/hooks/base-primitive/helpers/types";
 import {
   PRIMITIVE_PROPS,
   type PrimitiveProps,
   useBasePrimitiveProps,
-} from "../../utils/components/base-primitive/BasePrimitive";
-import {
-  getResponsiveProps,
-  getResponsiveValue,
-} from "../../utils/components/base-primitive/helpers/css";
-import type {
-  ResponsiveProp,
-  SpacingScale,
-} from "../../utils/components/base-primitive/helpers/types";
-import { Slot } from "../../utils/components/slot/Slot";
-import { cl } from "../../utils/helpers";
+} from "../../utils/hooks/base-primitive/useBasePrimitiveProps";
 import type { AsChildProps } from "../../utils/types";
 
 export type HGridProps = React.HTMLAttributes<HTMLDivElement> & {
