+++ @navikt/core/react/src/utils/hooks/base-primitive/helpers/css.ts
