+++ @navikt/core/react/src/utils/hooks/base-primitive/helpers/types.ts
